src/requirements.txt
quart
openai
python-dotenv
azure-identity
